requirements.txt
asgiref==3.8.1
Django==3.2
djangorestframework==3.12.4
flake8==6.0.0
flake8-isort==6.0.0
isort==5.13.2
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
